pyproject.toml
[tool.poetry]
name = "nonebot-plugin-nai3-bot"
version = "1.2.0"
description = "基于GPT4+NovelAI V3,Bot在将自然语言转换为NAI3提示词并绘图发送的同时，还会以自定义人格和你聊天。同时支持随机画风和随机同人图生成。"
authors = ["Alpaca <admin@alpaca.kim>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = "^2.2.1"
openai = "^1.30.1"
Pillow = "^9.1.0"
httpx = "^0.27.0"
numpy = "^1.25.2"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true 


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
